Merge from PatrickNiedzielski branch into unstable, from r222 to HEAD.  It compiles fine.
I needed to make two adjustments.  First, THROW is not in the API anymore, because it confliced on Windows.  I needed to change it to HUMMSTRUMM_THROW.  Also, in the X11 window file, I had to change an "Error" object to a new "Generic" object which I made.  Error cannot be used due to some internal API issues that I had to make.  Also, I forgot to add them to the repository...I need to do another checkin for that.